requirements.txt
# python==3.11.10
flask==3.1
tensorflow==2.19.0
pyyaml==6.0.2
polars==1.31.0
